Reorganise Select traits to not expose internal runtime types. Close #5160. Pending #8215.
