+++ src/libstd/std.rs
@@ -164,6 +164,7 @@ pub mod trie;
 
 pub mod task;
 pub mod comm;
+pub mod select;
 pub mod local_data;
